pyproject.toml
[project]
name = "claude-code-proxy"
version = "1.0.0"
description = "A simplified proxy for routing Anthropic requests to Bedrock/OpenRouter"
authors = [
    {name = "Claude Code Proxy Team"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "boto3>=1.34.0",
    "httpx>=0.25.2",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
            "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "uv>=0.1.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "bandit>=1.7.0",
]

[project.scripts]
claude-proxy = "main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "bandit>=1.7.0",
]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.flake8]
max-line-length = 100
ignore = ["E501", "W503"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["."]
python_files = ["test_*.py"] 